+++ tui/src/ui/components/lyric.rs
@@ -264,21 +264,20 @@ impl Model {
             }
 
             let mut line = String::new();
-            if song.lyric_frames_is_empty() {
-                self.lyric_set_lyric("No lyrics available.");
-                return;
-            }
 
             if let Some(l) = song.parsed_lyric() {
-                if l.unsynced_captions.is_empty() {
+                if l.captions.is_empty() {
                     self.lyric_set_lyric("No lyrics available.");
                     return;
                 }
                 if let Some(l) = l.get_text(self.time_pos) {
-                    line = l;
+                    line = l.to_string();
                 }
+            } else {
+                self.lyric_set_lyric("No lyrics available.");
+                return;
             }
-            self.lyric_set_lyric(&line);
+            self.lyric_set_lyric(line);
         }
     }